Prevent failure on legitimate buffers
The problem I've seen is the following:

Execute vim on a file, e.g. 'vim <some-c-file>', local_vimrc works
perfectly. Then, open another file of the same time in the same
directory using NERDTree - boom. It does not load the relevant
local_vimrc file.

I turned on verbosity and saw that the file is 'forbidden' because
the containing directory pathname was provided to
lh#project#is_eligible() which in turn used it on bufname(), which
returns an empty string because it matched the directory on more than
one buffer.

The underlying function lh#project#is_eligible() works better with
buffer IDs and not paths, I think that is what should be given to it.